Try though a fail-fast method to check table size.
This method is similar to format_singleline_table, except it fails early.
It formats a single field at a time, then checks the shape to see if we
are over budget. If so, we bail early.

This led to a performance regression of 4038.1% compared to naive
braces range: 44.26ms -> 1.8318s.
However, compared to original this is an improvement of 40% from 0.9.2 (3.0536s)
This is still clearly exponential, so should not be used